exercise_8/exercise 8/recommender-model/requirements.txt
flask
requests
pandas
scikit-surprise==1.1.1
joblib